pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "crisis-world"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "openenv-core>=0.2.3,<0.3",
    "pydantic>=2.0",
    "numpy>=1.24",
    "scipy>=1.11",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.4",
    "mypy>=1.10",
]
server = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.29",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[tool.ruff]
target-version = "py311"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.mypy]
python_version = "3.11"
strict = true
plugins = ["pydantic.mypy"]

[project.scripts]
# Server entry point - enables running via: uv run --project . server
# or: python -m CrisisWorld.server.app
server = "CrisisWorld.server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["CrisisWorld",
            "CrisisWorld.agents", 
            "CrisisWorld.cortex",
            "CrisisWorld.evaluation", 
            "CrisisWorld.protocols", 
            "CrisisWorld.schemas", 
            "CrisisWorld.server", 
            "CrisisWorld.tests"
        ]

[tool.setuptools.package-dir]
"CrisisWorld" = "."
"CrisisWorld.agents" = "agents"
"CrisisWorld.cortex" = "cortex"
"CrisisWorld.evaluation" = "evaluation"
"CrisisWorld.protocols" = "protocols"
"CrisisWorld.schemas" = "schemas"
"CrisisWorld.server" = "server"
"CrisisWorld.tests" = "tests"
